Bugfixing / Refactoring
