+++ plugins/org.projectforge.plugins.ihkexportnew/pom.xml
@@ -5,11 +5,11 @@
   <parent>
     <artifactId>org.projectforge.plugins</artifactId>
     <groupId>org.projectforge</groupId>
-    <version>6.25.0</version><!-- projectforge.version -->
+    <version>7.0-SNAPSHOT</version><!-- projectforge.version -->
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>org.projectforge.plugins.ihkexportNew</artifactId>
+  <artifactId>org.projectforge.plugins.ihkexportnew</artifactId>
   <packaging>jar</packaging>
 
   <dependencies>
