+++ plugins/org.projectforge.plugins.ihkexportnew/src/main/resources/META-INF/services/org.projectforge.framework.persistence.api.RightRightIdProviderService
@@ -1,2 +1,2 @@
-org.projectforge.plugins.ihkexportNew.IhkExportPluginUserRightIdNew$ProviderService
+org.projectforge.plugins.ihkexportnew.IhkExportPluginUserRightIdNew$ProviderService
