+++ plugins/org.projectforge.plugins.ihkexportold/pom.xml
@@ -5,11 +5,11 @@
   <parent>
     <artifactId>org.projectforge.plugins</artifactId>
     <groupId>org.projectforge</groupId>
-    <version>6.25.0</version><!-- projectforge.version -->
+    <version>7.0-SNAPSHOT</version><!-- projectforge.version -->
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>org.projectforge.plugins.ihkexportOld</artifactId>
+  <artifactId>org.projectforge.plugins.ihkexportold</artifactId>
   <packaging>jar</packaging>
 
   <dependencies>
