+++ plugins/org.projectforge.plugins.ihkexportold/src/main/java/org/projectforge/plugins/ihkexportold/IhkExportPluginOld.java
@@ -24,22 +24,22 @@
 package org.projectforge.plugins.ihkexportold;
 
 import org.projectforge.business.timesheet.TimesheetDao;
+import org.projectforge.menu.builder.MenuItemDef;
+import org.projectforge.menu.builder.MenuItemDefId;
 import org.projectforge.plugins.core.AbstractPlugin;
-import org.projectforge.web.MenuItemDef;
-import org.projectforge.web.MenuItemDefId;
 import org.projectforge.web.plugin.PluginWicketRegistrationService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 /**
  * Your plugin initialization. Register all your components such as i18n files, data-access object etc.
- * 
+ *
  * @author Kai Reinhard (k.reinhard@micromata.de)
  */
 @Component
 public class IhkExportPluginOld extends AbstractPlugin
 {
-  public static final String ID = "ihkexportOld";
+  public static final String ID = "ihkexportold";
 
   public static final String RESOURCE_BUNDLE_NAME = "IhkExportOldI18nResources";
 
@@ -54,15 +54,14 @@ public class IhkExportPluginOld extends AbstractPlugin
   {
 
     // Register it:
-    register(ID, TimesheetDao.class, ihkExportDao, "plugins.ihkexportOld12345");
+    register(ID, TimesheetDao.class, ihkExportDao, "plugins.ihkexportold");
 
     // Register the web part:
     pluginWicketRegistrationService.registerWeb(ID);
 
     // Register the menu entry as sub menu entry of the misc menu:
-    final MenuItemDef parentMenu = pluginWicketRegistrationService.getMenuItemDef(MenuItemDefId.MISC);
-    pluginWicketRegistrationService
-        .registerMenuItem(new MenuItemDef(parentMenu, ID, 10, "plugins.ihkexportOld.menu", IhkExportPageOld.class));
+    pluginWicketRegistrationService.registerMenuItem(MenuItemDefId.MISC, MenuItemDef.create(ID, "plugins.ihkexportOld.menu"),
+            IhkExportPageOld.class);
 
     // Define the access management:
     registerRight(new IhkExportRightOld(accessChecker));
