+++ plugins/org.projectforge.plugins.ihkexportold/src/main/resources/META-INF/services/org.projectforge.framework.persistence.api.RightRightIdProviderService
@@ -1,2 +1,2 @@
-org.projectforge.plugins.ihkexportOld.IhkExportPluginUserRightIdOld$ProviderService
+org.projectforge.plugins.ihkexportold.IhkExportPluginUserRightIdOld$ProviderService
