+++ plugins/org.projectforge.plugins.ihkexportold/src/main/resources/META-INF/services/org.projectforge.plugins.core.ProjectforgePluginService
@@ -1,1 +1,1 @@
-org.projectforge.plugins.ihkexportOld.IhkExportProjectforgePluginServiceOld
+org.projectforge.plugins.ihkexportold.IhkExportProjectforgePluginServiceOld
